inference-function/requirements.txt
onnxruntime >= 1.0.0, < 2.0.0
Pillow == 7.0.0
numpy == 1.18.1
